Add generic async loader for loading non-react modules
